requirements.txt
streamlit
langchain
pypdf
chromadb
sentence-transformers
groq
python-dotenv
numpy